Refactor: tools,libpacemaker: Move profile_one and profile_all.
These functions do one of the major tasks of crm_simulate - generate
profiling data.  Moving them out into libpacemaker provides a way for
other tools to do the same thing, if needed.  Also, rename them to make
it more clear exactly what they do.
