+++ include/pcmki/pcmki_simulate.h
@@ -14,6 +14,39 @@
 #include <stdbool.h>
 
 /**
+ * \brief Profile the configuration updates and scheduler actions in a single
+ *        CIB file, printing the profiling timings.
+ *
+ * \note \p data_set->priv must have been set to a valid \p pcmk__output_t
+ *       object before this function is called.
+ *
+ * \param[in] xml_file The CIB file to profile.
+ * \param[in] repeat   Number of times to run.
+ * \param[in] data_set Working set for the cluster.
+ * \param[in] use_date The date to set the cluster's time to (may
+ *                     be NULL).
+ */
+void pcmk__profile_file(const char *xml_file, long long repeat, pe_working_set_t *data_set,
+                        char *use_date);
+
+/**
+ * \brief Profile the configuration updates and scheduler actions in every
+ *        CIB file in a given directory, printing the profiling timings for
+ *        each.
+ *
+ * \note \p data_set->priv must have been set to a valid \p pcmk__output_t
+ *       object before this function is called.
+ *
+ * \param[in] dir      A directory full of CIB files to be profiled.
+ * \param[in] repeat   Number of times to run on each input file.
+ * \param[in] data_set Working set for the cluster.
+ * \param[in] use_date The date to set the cluster's time to (may
+ *                     be NULL).
+ */
+void pcmk__profile_dir(const char *dir, long long repeat, pe_working_set_t *data_set,
+                       char *use_date);
+
+/**
  * \brief Set the date of the cluster, either to the value given by
  *        \p use_date, or to the "execution-date" value in the CIB.
  *
