+++ tools/crm_simulate.c
@@ -652,85 +652,6 @@ setup_input(const char *input, const char *output, GError **error)
     }
 }
 
-static void
-profile_one(const char *xml_file, long long repeat, pe_working_set_t *data_set, char *use_date)
-{
-    pcmk__output_t *out = data_set->priv;
-    xmlNode *cib_object = NULL;
-    clock_t start = 0;
-    clock_t end;
-
-    cib_object = filename2xml(xml_file);
-    start = clock();
-
-    if (get_object_root(XML_CIB_TAG_STATUS, cib_object) == NULL) {
-        create_xml_node(cib_object, XML_CIB_TAG_STATUS);
-    }
-
-
-    if (cli_config_update(&cib_object, NULL, FALSE) == FALSE) {
-        free_xml(cib_object);
-        return;
-    }
-
-    if (validate_xml(cib_object, NULL, FALSE) != TRUE) {
-        free_xml(cib_object);
-        return;
-    }
-
-    for (int i = 0; i < repeat; ++i) {
-        xmlNode *input = (repeat == 1)? cib_object : copy_xml(cib_object);
-
-        data_set->input = input;
-        pcmk__set_effective_date(data_set, false, use_date);
-        pcmk__schedule_actions(data_set, input, NULL);
-        pe_reset_working_set(data_set);
-    }
-
-    end = clock();
-    out->message(out, "profile", xml_file, start, end);
-}
-
-#ifndef FILENAME_MAX
-#  define FILENAME_MAX 512
-#endif
-
-static void
-profile_all(const char *dir, long long repeat, pe_working_set_t *data_set, char *use_date)
-{
-    pcmk__output_t *out = data_set->priv;
-    struct dirent **namelist;
-
-    int file_num = scandir(dir, &namelist, 0, alphasort);
-
-    if (file_num > 0) {
-        struct stat prop;
-        char buffer[FILENAME_MAX];
-
-        out->begin_list(out, NULL, NULL, "Timings");
-
-        while (file_num--) {
-            if ('.' == namelist[file_num]->d_name[0]) {
-                free(namelist[file_num]);
-                continue;
-
-            } else if (!pcmk__ends_with_ext(namelist[file_num]->d_name,
-                                            ".xml")) {
-                free(namelist[file_num]);
-                continue;
-            }
-            snprintf(buffer, sizeof(buffer), "%s/%s", dir, namelist[file_num]->d_name);
-            if (stat(buffer, &prop) == 0 && S_ISREG(prop.st_mode)) {
-                profile_one(buffer, repeat, data_set, use_date);
-            }
-            free(namelist[file_num]);
-        }
-        free(namelist);
-
-        out->end_list(out);
-    }
-}
-
 static GOptionContext *
 build_arg_context(pcmk__common_args_t *args, GOptionGroup **group) {
     GOptionContext *context = NULL;
@@ -852,7 +773,7 @@ main(int argc, char **argv)
 
     if (options.test_dir != NULL) {
         data_set->priv = out;
-        profile_all(options.test_dir, options.repeat, data_set, options.use_date);
+        pcmk__profile_dir(options.test_dir, options.repeat, data_set, options.use_date);
         rc = pcmk_rc_ok;
         goto done;
     }
